requirements.txt
requests
python-dotenv
fastapi
uvicorn
pydantic
openai
streamlit
